+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.4.2] - 2025-09-22
+
+### Added
+
+- 42 new daily puzzles.
+
 ### Fixed
 
 - Changed start state of "End Zone" puzzle so that it is solvable.